requirements.txt
numpy==1.24.3
pandas==2.0.2
oandapyV20==0.7.2
scikit-learn==1.2.2
matplotlib==3.7.1
python-dotenv==1.0.0
aiohttp==3.8.4
